requirements.txt
Django>=5.0,<6.0
speedtest-cli==2.1.3
scapy==2.5.0
matplotlib==3.7.1
psutil==5.9.5
numpy==1.24.3
ping3==4.0.4
